Add a helper script to add new entries
This helper script is meant to be run by automation. It takes two
parameters, via `--date=<date>` and `--commit=<commit>`, and expects the
remainder of the command-line arguments to be URLs pointing to snapshot
artifacts.

This script doubles as a (CommonJS) module: While it can be run via the
command-line, it can also be used e.g. in an `actions/github-script`
step via `const main = require('./add-entry.js'); main(...);`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>