Backfill the missing Git installers
I opened a PR at
https://github.com/git-for-windows/git-for-windows-automation/pull/112
to fix the problem where the recent snapshots were missing the
installers.

This addresses the problem for new snapshots, for existing ones I ran
https://github.com/git-for-windows/git-snapshots/actions/runs/13264619184
to upload the missing installers.

This here commit now updates the `index.html` file to add the
previously-missing links to these previously-missing installers.

This trick was performed by duplicating the respective `PortableGit`
lines and then running the following two substitutions in vim:

  :s/Portable\([^"]*\)\.7z\.exe/\1.exe/g

  :s/>Portable Git (self-extracting <tt>.7z<\/tt> archive): </>Git for Windows installer: <

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>