add-entry.js: optionally allow "back-filling" missing releases
There has been a problem with the `upload-snapshot` workflow, fixed in
https://github.com/git-for-windows/gfw-helper-github-app/pull/128, that
prevented the v2.49.0-rc* releases from being added.

I manually ran the automation for -rc2, but for -rc0 and -rc1 I want to
use a different approach: I want to cross-link to the existing releases
in `git-for-windows/git`.

The new `--backfill-release=<tag-name>` option allows just that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>